+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Add GCS support to S3 backend
 ### Changed
+- Fixed docs refering to github in s3 backend
 ### Removed
 
 ## [0.20.0]